+++ packages/cli/src/lib/system/index.ts
@@ -1,5 +1,6 @@
 export * from "./child-process";
 export * from "./docker";
+export * from "./env";
 export * from "./file-lock";
 export * from "./file-watcher";
 export * from "./filesystem";
